+++ editor/src/components/canvas/canvas-strategies/keyboard-interaction.test-utils.tsx
@@ -66,24 +66,3 @@ export function pressKeys(
 
   return finalEditor
 }
-
-export const shiftModifier: Modifiers = {
-  alt: false,
-  cmd: false,
-  ctrl: false,
-  shift: true,
-}
-
-export const cmdModifier: Modifiers = {
-  alt: false,
-  cmd: true,
-  ctrl: false,
-  shift: false,
-}
-
-export const shiftCmdModifier: Modifiers = {
-  alt: false,
-  cmd: true,
-  ctrl: false,
-  shift: true,
-}
