+++ editor/src/utils/modifiers.ts
@@ -14,12 +14,26 @@ export const emptyModifiers: Modifiers = optionalDeepFreeze({
   shift: false,
 })
 
-export const shiftModifier: Modifiers = optionalDeepFreeze({
+export const shiftModifier: Modifiers = {
   alt: false,
   cmd: false,
   ctrl: false,
   shift: true,
-})
+}
+
+export const cmdModifier: Modifiers = {
+  alt: false,
+  cmd: true,
+  ctrl: false,
+  shift: false,
+}
+
+export const shiftCmdModifier: Modifiers = {
+  alt: false,
+  cmd: true,
+  ctrl: false,
+  shift: true,
+}
 
 export const Modifier = {
   modifiersForKeyboardEvent: function (event: KeyboardEvent): Modifiers {
